requirements.txt
# Core dependencies
langchain
langchain-openai
langgraph
openai
langchain-community
langchain-core
langchain-experimental
# Data processing
pandas
numpy
openpyxl  # For Excel file support

# Data visualization
plotly
matplotlib
seaborn

# Web interface
streamlit
streamlit-chat

# Document processing
pytesseract  # OCR for images
pdf2image    # Convert PDF to images
PyPDF2       # PDF processing
pillow       # Image processing
opencv-python-headless  # Image processing without GUI
google-generativeai
# HTTP requests for API calls
requests

# Utilities
python-dotenv  # Environment variable management
tqdm           # Progress bars
pydantic       # Data validation
python-multipart
fastapi
uvicorn
tabulate
pandasai